requirements.txt
firecrawl-py>=0.0.16
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.0.0
rich>=13.0.0
python-slugify>=8.0.0
python-dateutil>=2.8.2
PyYAML>=6.0.0